refactor: extract grid helpers to separate module
Move signThumbnailUrl() and buildGridPanelData() from crud.ts to
new grid-helpers.ts module for better separation of concerns.

- Creates packages/cms/grid-helpers.ts with grid-specific functions
- Removes ~210 lines of grid logic from the main CRUD handlers
- Keeps getSafeReturnUrl() in crud.ts (used by multiple handlers)
- No functional changes, all 1208 tests pass
